pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kql-mcp"
version = "0.1.0"
description = "The best MCP server for KQL (Kusto Query Language) — ADX, Log Analytics, and Sentinel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.6.0",
    "azure-kusto-data>=4.4.0",
    "azure-monitor-query>=1.4.0",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.10.0",
    "pandas>=2.2.0",
    "tabulate>=0.9.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
]

[project.scripts]
kql-mcp = "kql_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kql_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
